refactor: improve quality by renaming some folders
