+++ frontend/projects/webapp/src/app/feature/budget/budget-details/data-core/budget-item-data-builder.ts
@@ -1,6 +1,6 @@
 import type { Transaction, BudgetLine, TransactionKind } from 'pulpe-shared';
 import { calculateAllConsumptions } from '@core/budget/budget-line-consumption';
-import { isRolloverLine } from '@core/rollover/rollover-types';
+import { isRolloverLine } from '@core/budget/rollover/rollover-types';
 import type {
   BudgetLineTableItem,
   GroupHeaderTableItem,
