+++ frontend/projects/webapp/src/app/feature/budget/budget-details/store/budget-details-store.ts
@@ -9,7 +9,7 @@ import {
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { BudgetApi } from '@core/budget/budget-api';
 import { Logger } from '@core/logging/logger';
-import { createRolloverLine } from '@core/rollover/rollover-types';
+import { createRolloverLine } from '@core/budget/rollover/rollover-types';
 import { StorageService } from '@core/storage/storage.service';
 import { STORAGE_KEYS } from '@core/storage/storage-keys';
 import { TransactionApi } from '@core/transaction/transaction-api';
