+++ frontend/projects/webapp/src/app/feature/budget/budget-list/create-budget/services/template-store.ts
@@ -1,7 +1,7 @@
 import { Injectable, computed, inject, signal } from '@angular/core';
 import { firstValueFrom } from 'rxjs';
 import { type TemplateLine, type BudgetTemplate } from 'pulpe-shared';
-import { TemplateApi } from '@core/template/template-api';
+import { TemplateApi } from '@core/budget-template/template-api';
 import { Logger } from '@core/logging/logger';
 import {
   TemplateTotalsCalculator,
