+++ frontend/projects/webapp/src/app/feature/complete-profile/complete-profile-store.ts
@@ -1,5 +1,5 @@
 import { Injectable, inject, signal, computed } from '@angular/core';
-import { ProfileSetupService, type ProfileData } from '@core/profile';
+import { ProfileSetupService, type ProfileData } from '@core/complete-profile';
 import { BudgetApi } from '@core/budget';
 import { Logger } from '@core/logging/logger';
 import { PostHogService } from '@core/analytics/posthog';
