+++ frontend/projects/webapp/src/app/feature/current-month/utils/financial-entry-mapper.ts
@@ -1,4 +1,4 @@
-import { isRolloverLine } from '@core/rollover/rollover-types';
+import { isRolloverLine } from '@core/budget/rollover/rollover-types';
 import type { BudgetLine, Transaction } from 'pulpe-shared';
 import type { FinancialEntryModel } from '../models/financial-entry.model';
